requirements.txt
mcp[cli]
requests>=2.28
python-dotenv>=1.0